+++ apps/typegpu-docs/src/pages/new/index.astro
@@ -120,7 +120,7 @@ import { HeroEffect } from '../../components/hero-effect/HeroEffect';
           TypeGPU gives you the tools to build anything WebGPU can do.
         </p>
       </div>
-      <LandingExampleGrid client:visible />
+      <LandingExampleGrid client:only="react" />
     </section>
 
     <!-- Code Showcase -->
